Add number functionality and start operator func
Add responsive buttons, which look like they are being clicked, and
update the screen when they are clicked. Also started to work on
operator button functionality, namely plus and equals. For whatever
reason, the operate function works in the console, but when returning
a result to the screen, it results in concatenation instead of
summation. (9+8=98, instead of 17). That's my next bug to work out,
tomorrow.