Fix operator logic and display issues
Instead of using integers the whole time, the solution I came up with
was to declare empty strings for the two argument variables and the
operator variable. Then when the input is passed through the input
function to the operate function, `parseInt()` is used to convert the
arg strings back into integers. So now the calculator works with whole
variables, and can return results with decimal values. However, decimal
values are automatically disregarded when passed back through the
operate function (not to mention the decimal and pos/neg buttons don't
have functionality... yet.). So yeah, now it works for simple
calculations buttons, which means there is more work to do!