Clean up code a little
Cleaned up the old commented out code which is now unneeded. This
project now satisfies the basic specs for the assignment, though it
still has ample bugs. For example if you put in more input beyond
the screens capacity, it will still store the input in the variables
without displaying it. I also have yet to add keyboard functionality,
which I would like to come back and add if I decide to use this for a
portfolio piece in the future. For now I am satisfied and looking
forward to progressing with the course.