Remove the '=' reset functionality
Although the project spec asks more input to start a new operation from
scratch, I started to miss the old functionality where if you wanted to
perform more operations on the solution to the previous operation. So,
I changed it back to how it was before. It feels better now.